Cleanup, bug fix, and addition of reusing functions.
This commit adds several things.

First off all, the reuse file is no used to update the rawspec after parsing.
Therefor the core modules export all of their new functions.

Secondly, there were some severe bugs with the reuse code of functions.
The wrong module was imported and the wrong function (that for enums) 
was used to query the location.

Thirdly the code has been cleanup, out-commented code was removed.