Changes some parts to a less powerful monad and repackages.
The monad used for making the modules is switched from the full package
builder monad to one that has only access to the module that is being build.
This is done by changing builder to build a Module in stead of a package and
use the newly defined RawPBuilder for building the package.

Additionaly the 'addModuleAndActivate' function is moved to the package 
CodeGeneration.