Restructures and changes the RawSpec/SpecValue use Monoid.
The restructure is done to split up the RawSpec in several parts, one for the
defenitions, one for the quering of the spec and one for processing it to a
'generatable' form.

Furthermore a Monoid instance is added for RawSpec. This Monoid instance
is used to simplify generating the RawSpec to merging several smaller ones.