Adds Sorting of the imports.
This implements the sorting function which changes all the defenitions in
order to prevent import cycles. This is done by using the first category that
is passed for a value to be the place where it's defined. This works as the 
traversing order traverses category for category. 
An import cycle may still ocure due to the aliasing of enumerations. This 
remains to be solved