Stores the originals function name for the ffi import.
The name that was found in the RawSpec is assumed to be the name of the 
function to import. This is not correct as this name could be changed and 
thus also the function imported. To prevent this the name that should be 
imported is stored with the function definition.