Refactors the string names to a type famely in SpecValue.
This refactiring is done to prevent the use of Strings, as the use of Strings
keeps too much posibility for unwanted edits on the names. Thus in effect
this tries to add another layer of type safety. This is done by only allowing
`ValueName`s to be constructed from Strings (for parsing) and let the
deconstruction go to the Name type (from src-exts). Assuming the use
of `unname` or simmilar functions is done with great caution this should
bring another layer of safety.

The renname function that was in Spec.Processing is also removed as this
function is not used anymore.