Replaces the PackageBuilder (fix deprecate merge)
The Package builder monad from Code.Generating will probably be moved
or deleted in the future. Furthermore as it is old code from a period
with less experience I don't like it any more. Therefore it is replaced
by a Writer monad over lists of modules (with some extra information),
which are all defined in Code.Builder.

As a result of the rewrite several differences showed up between the
reference and the output. This was due to merging of the deprecated
and normal versions of category modules. This is fixed by adding
"Compatibility" to the modules with deprecated contents.