Adds -Werror, changes -fno-warns
There shouldn't be any duplicate exports after the previous fix, so that
warning is reabled. The unused-imports is needed for some modules where
Foreig.C.Types is imported but not used as it only defines functions of
type IO (). These warnings can be (I think) safely ignored. All the
other warnings indicate that something has gone wrong and should
therefore be treated as errors.