Adds warnings to the generated files.
A warning message can be inserted between the pragmas and the module,
which is used to warn against editing. The default is not to add any
text, but this has been overriden in testrun.sh
Simple haskell warnings and deprecation texts can be added, this is
used in the compatibility modules, the default is on. (Fixes #5)