Refactors RawSpec merging Enum and Func into one.
The previous Maps that form the Spec made a strict difference between
enumerations and functions. This caused quite some code duplication and
did only provide a little extra safety. In addition it made quite some
functions a bit harder to read.Therefore this experiment on merging the
Enum and Func related types into one.

The result is that all the Mappings are now simple Maps without the
extra duality. The resulting code is less duplicated but has still two
more problematic points. Reusing other definitions is only used for
enumerations but it is now also implemented for function. It might need
a bit of tweaking. In addition the name still contains a separate tag
which indicates if corresponds to a function or enumeration, which can
probably be made more type safe.